Add and Consolidate Formatters
