+++ components/AnalysisSummary.test.tsx
@@ -228,7 +228,7 @@ test('renders the full tables with some analyses', () => {
           <strong>
             Last analyzed: 
           </strong>
-          2020-05-10
+          2020-05-10T00:00:00.000+00:00
         </div>
         <div
           class="MuiPaper-root MuiTableContainer-root MuiPaper-elevation1 MuiPaper-rounded"
@@ -501,7 +501,7 @@ test('renders the full tables with some analyses', () => {
           <strong>
             Last analyzed: 
           </strong>
-          2020-05-10
+          2020-05-10T00:00:00.000+00:00
         </div>
         <div
           class="MuiPaper-root MuiTableContainer-root MuiPaper-elevation1 MuiPaper-rounded"
